+++ examples/remix/app/routes/complete.tsx
@@ -1,7 +1,10 @@
 import {json, type ActionFunctionArgs} from '@remix-run/node';
 import {Copilot, type CompletionRequestBody} from 'monacopilot';
 
-const copilot = new Copilot(process.env.ANTHROPIC_API_KEY!);
+const copilot = new Copilot(process.env.GROQ_API_KEY!, {
+  provider: 'groq',
+  model: 'llama-3-70b',
+});
 
 export const action = async ({request}: ActionFunctionArgs) => {
   const body: CompletionRequestBody = await request.json();