+++ CHANGELOG.md
@@ -1,3 +1,36 @@
+
+
+## [0.15.13](https://github.com/arshad-yaseen/monacopilot/compare/v0.15.12...v0.15.13) (2024-12-22)
+
+
+### 🔧 CI/CD
+
+* close inactve issues workflow ([4ef021f](https://github.com/arshad-yaseen/monacopilot/commit/4ef021fac01c3a163e4797d60a9baa030323d91d))
+
+## [0.15.12](https://github.com/arshad-yaseen/monacopilot/compare/v0.15.11...v0.15.12) (2024-12-22)
+
+### 🔧 Maintenance
+
+- require explicit model specification, remove default model ([1726beb](https://github.com/arshad-yaseen/monacopilot/commit/1726beb29e1e4ee0a8833428d0782882ebaf4136))
+
+## [0.15.11](https://github.com/arshad-yaseen/monacopilot/compare/v0.15.10...v0.15.11) (2024-12-21)
+
+### 🔧 Maintenance
+
+- minimal prompt ([a7101c6](https://github.com/arshad-yaseen/monacopilot/commit/a7101c6769961ab337bb3e792f50ba4422026d22))
+
+## [0.15.10](https://github.com/arshad-yaseen/monacopilot/compare/v0.15.9...v0.15.10) (2024-12-21)
+
+### 🔧 Maintenance
+
+- improve prompt ([6cf6223](https://github.com/arshad-yaseen/monacopilot/commit/6cf62231fe7b42b9d555dd87cddce197901cad9b))
+
+## [0.15.9](https://github.com/arshad-yaseen/monacopilot/compare/v0.15.8...v0.15.9) (2024-12-21)
+
+### 🔧 Maintenance
+
+- add fetch timeout ([3df46dd](https://github.com/arshad-yaseen/monacopilot/commit/3df46dd74cd65f3d2ae85d5a288d21b7d0969870))
+
 ## [0.15.8](https://github.com/arshad-yaseen/monacopilot/compare/v0.15.7...v0.15.8) (2024-12-15)
 
 ### 📚 Documentation